bot/requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b3
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.5.18
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
magic-filter==1.0.7
multidict==6.0.2
pydantic==1.9.0
Pygments==2.12.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2
